Add a unuque identifier for installed packages (part 2 of 9)
Note: this patch doesn't build on its own, you also need the rest of
the patch series.

Compatibility with older GHCs.  When reading a package database
created by an older version of GHC without installedPackageIds, we
have to fake an InstalledPackageId for the internal
InstalledPackageInfo.

So, when reading in an InstalledPackageInfo from an older GHC, we set
the installedPackageId to be the textual representation of the
PackageIdentifier: i.e. <package>-<version>.  Now, previously the
depends field of InstalledPackageInfo was [PackageIdentifier], and is
now [InstalledPackageId], so we will read each PackageIdentifier as an
InstalledPackageId (a String).  The dependencies will still point to
the correct package, however, because we have chosen the
installedPackageId to be the textual representation of the
PackageIdentifier.